Fix config for Audio
