+++ ios/flare/AudioSensor.m
@@ -15,4 +15,10 @@ RCT_EXTERN_METHOD(
   rejecter: (RCTPromiseRejectBlock)reject
 )
 
+// Expose focus method
+RCT_EXTERN_METHOD(
+  focus: (RCTPromiseResolveBlock)resolve
+  rejecter: (RCTPromiseRejectBlock)reject
+)
+
 @end
