+++ src/App.tsx
@@ -20,16 +20,6 @@ Sentry.init({
   enableAutoSessionTracking: true,
 })
 
-Audio.setAudioModeAsync({
-  allowsRecordingIOS: false,
-  staysActiveInBackground: true,
-  interruptionModeIOS: Audio.INTERRUPTION_MODE_IOS_DO_NOT_MIX,
-  playsInSilentModeIOS: true,
-  shouldDuckAndroid: true,
-  interruptionModeAndroid: Audio.INTERRUPTION_MODE_ANDROID_DO_NOT_MIX,
-  playThroughEarpieceAndroid: false,
-})
-
 // Base container for all screens
 export default function App() {
   const [loaded, setLoaded] = useState(false)
